backend/pyproject.toml
[project]
name = "hackuta-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn>=0.30.0",
    "google-genai>=0.8.0",
]
